aarch64: fix FAR_EL1 width in writeFAR
Fixes a bug where writeFAR was only writing to the low 32 bits of the
64-bit FAR_EL1 register.

+++ include/arch/arm/armv/armv8-a/64/armv/vcpu.h
@@ -256,7 +256,7 @@ static inline word_t readFAR(void)
 
 static inline void writeFAR(word_t reg)
 {
-    MSR(REG_FAR_EL1, (uint32_t)reg);
+    MSR(REG_FAR_EL1, reg);
 }
 
 /* ISR is read-only */
